+++ ctts/tests/test_prediction_api.py
@@ -1,6 +1,6 @@
 from fastapi.testclient import TestClient
 
-from ..api import app,prefix
+from ..api import app, prefix
 
 client = TestClient(app)
